move all qd_suite tasks
